[Refactoring] Move the action listener from fragment to viewmodel
MarkFragment
MarkListAdapter
MarkViewModel
